feat(membership): close ban/suspend revocation, course gates, i64 cents SSOT
Phase A (security gaps closed):
- AuthUser + OptionalAuthUser re-check banned_at/suspended_at on every
  authed request. Pre-fix: banned users kept access for up to 24h
  (JWT_EXPIRATION_HOURS) until their token expired. Post-fix:
  revocation is immediate. Lazy-unsuspend logic preserved.
- New PaidUser extractor for subscription-gated handlers (Active or
  Trialing only; admin bypass).
- Course enrollment gate (POST /api/member/courses/{id}/enroll):
  free → any authed; sub-included → requires Active/Trialing;
  pay-per-course → 403 (purchase ledger not yet built).
- Lesson content redaction in GET /api/courses/{slug}: non-preview
  lesson body / video_url stripped for unauthorized viewers.
- subscriptions.canceled_at now populated on customer.subscription.deleted
  webhook (COALESCE-protected against Stripe retries).

Migration 081 (i64 cents SSOT): widens 7 money columns from INTEGER
(max ~$21.4M) to BIGINT (max ~$92 quintillion). 13 Rust struct fields
flip from i32 → i64 in lockstep. SubscriptionStatus + SubscriptionPlan
enums gain Copy + Eq. OpenAPI snapshot + frontend types regenerated.

Phase B (43 new integration tests, all green):
- lifecycle_revocation.rs (12): ban / suspend / expired-suspension
  self-heal / hard-delete / admin ban / future-dated ban / OptionalAuth
  leakage / refresh post-ban / un-ban+un-suspend symmetry.
- subscription_lifecycle_access.rs (13): per-status is_active matrix
  (Active/Trialing/PastDue/Unpaid/Canceled/Paused/no-row); transitions;
  canceled_at populated + COALESCE-preserved across replays.
- course_access_gates.rs (18): free / sub-included / pay-per-course
  enrollment + content redaction matrix; preview-lesson visibility;
  admin bypass; unpublished-course 404.

Verification:
- cargo fmt --check + cargo clippy --all-targets -D warnings: clean
- cargo test --lib: 524 pass / 0 ignored
- cargo test --tests: 893 pass / 0 ignored / 0 failed (43 binaries)
- pnpm lint + pnpm check + pnpm test:unit: clean

Phase C (real Stripe test-mode E2E with stripe CLI) is the follow-up;
blocked on operator pasting real sk_test_* keys into local .env files.

+++ backend/src/handlers/pricing.rs
@@ -734,7 +734,7 @@ pub(crate) struct PriceProtectionRequest {
     /// `false` = remove protection (subscription will receive price updates).
     pub enabled: bool,
     /// Optional: lock in a specific grandfathered amount for display purposes.
-    pub grandfathered_price_cents: Option<i32>,
+    pub grandfathered_price_cents: Option<i64>,
     pub grandfathered_currency: Option<String>,
 }
